+++ app/src/main/java/com/group15/fitnesstracker/onboarding/login/LoginPresenter.kt
@@ -33,7 +33,7 @@ class LoginPresenter(val view: LoginContract.View,
 
                     sharedPref?.edit()
                             ?.putBoolean(Constants.USER_LOGGED_IN, true)
-                            ?.putInt(Constants.CURRENT_USER_ID, it.id)
+                            ?.putLong(Constants.CURRENT_USER_ID, it.id)
                             ?.apply()
